% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_viz.R
\name{prepare10x_from_SpatialExperiment}
\alias{prepare10x_from_SpatialExperiment}
\title{Title Prepares Data For SpatialView Visualization from SpatialExperiment}
\usage{
prepare10x_from_SpatialExperiment(
  speObj,
  data.paths,
  export.path,
  projectName = "spatial",
  clusterCol = "label",
  clusterNames = NULL,
  clusterGenes = NULL,
  clustColors = NULL,
  slot = "data",
  assayName = "logcounts",
  multiSamplePattern = "_",
  exprRound = 3,
  spatialSubDir = "outs/spatial",
  sampleInfo = NULL,
  downloadRepo = TRUE,
  spatialviewRepo = "https://github.com/kendziorski-lab/spatialview/archive/refs/tags/",
  spatialviewVersion = "spatialview-latest",
  port = 8878,
  launchApp = TRUE,
  export_sparse = TRUE,
  data_file_name_expressions = "expression_matrix.csv",
  data_file_name_expressions_sparse = "expression_matrix_sparse.txt",
  data_file_name_genes = "genes.csv",
  data_file_name_barcodes = "barcodes.csv",
  verbose = FALSE,
  config.list = NA
)
}
\arguments{
\item{speObj}{SpatialExperiment object.}

\item{data.paths}{String array of paths for the raw samples.
Number of elements in the array should same as number of samples used in the SpatialExperiment object.}

\item{export.path}{Directory path for exporting the files.
If downloadRepo = TRUE, then the SpatialView application will be created at this place.
If downloadRepo = FALSE, then download SpatialView from GitHub manually and set export.path to SpatialView/data.}

\item{projectName}{Name of the project.
If downloadRepo= TRUE, the projectName is used to create the peoject directory at export.path, else ignores.}

\item{clusterCol}{Cluster column name in the SpatialExperiment object metadata.}

\item{clusterNames}{Array of names for clusters matching to clusters in 'clusterCol'.}

\item{clusterGenes}{List of arrays for genes specific to clusters, if interested to see in the visualization.}

\item{clustColors}{Array of hex colors for clusters, else default colors will be used.}

\item{slot}{Slot name in speObject for extracting data. Default 'data'.}

\item{assayName}{Assay name in speObject. If not provided then default Assay in speObject will be used.}

\item{multiSamplePattern}{Separator used for multiple samples in speObject. Default '_'.}

\item{exprRound}{The expressions are rounded for effectively saving space. Default rounding will be three decimal places.}

\item{spatialSubDir}{Sub-directory name where spatial files are located for 10x.}

\item{sampleInfo}{A dataframe with metadata information for the samples. Rows are for samples.}

\item{downloadRepo}{If TRUE (default), downloads SpatialView from GitHub repository, and runs the SpatialView application in local computer.}

\item{spatialviewRepo}{If downloadRepo = TRUE, then downloads the SpatialView from GitHub repository.}

\item{spatialviewVersion}{If downloadRepo = TRUE, then downloads the specified version of SpatialView.}

\item{port}{Port to be used for running SpatialView. If downloadRepo=FALSE, it's ignored.}

\item{launchApp}{If TRUE (default), then launches the web application when downloadRepo is TRUE, else ignores.}

\item{export_sparse}{If TRUE (default, recommended), sparse expression matrix will be exported, else dense csv format will be used.}

\item{data_file_name_expressions}{Name of the exported expression matrix in csv format. Ignores if export_sparse = TRUE.}

\item{data_file_name_expressions_sparse}{Name of the exported expression matrix in sparse format. Ignores if export_sparse = FALSE.}

\item{data_file_name_genes}{Name of the exported file containing names of the genes. Ignores if export_sparse = FALSE.}

\item{data_file_name_barcodes}{Name of the exported file containing barcodes. Ignores if export_sparse = FALSE.}

\item{verbose}{TRUE/FALSE whether to print comments.}

\item{config.list}{A list of named strings for changing the configuration.
This works if 'downloadRepo' is set to TRUE.
Configuration changes are having low precedence as compared to other parameters.}
}
\value{
None
}
\description{
Title Prepares Data For SpatialView Visualization from SpatialExperiment
}
