% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_viz.R
\name{prepare10xVisium_from_seurat}
\alias{prepare10xVisium_from_seurat}
\title{Prepares Data from Seurat object and launches SpatialView web application locally.}
\usage{
prepare10xVisium_from_seurat(
  seuratObj,
  dataPaths,
  exportPath,
  projectName = "spatial",
  clusterColumn = "seurat_clusters",
  clusterNames = NULL,
  clusterGenes = NULL,
  clusterColors = NULL,
  dimPlot = "umap",
  slot = "data",
  assayName = NULL,
  multiSamplePattern = "_",
  exprRound = 3,
  spatialSubDir = "spatial",
  sampleInfo = NULL,
  downloadRepo = TRUE,
  spatialviewRepo = "https://github.com/kendziorski-lab/spatialview/archive/refs/tags/",
  spatialviewVersion = "spatialview-latest",
  port = 8878,
  launchApp = TRUE,
  export_sparse = TRUE,
  data_file_name_expressions = "expression_matrix.csv",
  data_file_name_expressions_sparse = "expression_matrix_sparse.txt",
  data_file_name_genes = "genes.csv",
  data_file_name_barcodes = "barcodes.csv",
  configList = NA,
  verbose = FALSE
)
}
\arguments{
\item{seuratObj}{Seurat object.}

\item{dataPaths}{String array of paths for the raw samples.
Number of elements in the array should same as number of samples used in the Seurat object.}

\item{exportPath}{Directory path for exporting the files.
If downloadRepo = TRUE (default), then the SpatialView application will be created at this place.
If downloadRepo = FALSE, then download SpatialView from GitHub manually and set exportPath to SpatialView/data.}

\item{projectName}{Name of the project.
If downloadRepo= TRUE (default), the projectName is used to create the peoject directory at exportPath, else ignores.}

\item{clusterColumn}{Cluster column name in the Seurat object metadata.}

\item{clusterNames}{Array of names for clusters matching to clusters in 'clusterColumn'.}

\item{clusterGenes}{List of arrays for genes specific to clusters, if interested to see in the visualization.}

\item{clusterColors}{Array of hex colors for clusters, else default colors will be used.}

\item{dimPlot}{Reduced dimension to be used for cluster visualization.}

\item{slot}{Slot name in seuratObject for extracting data. Default 'data'.}

\item{assayName}{Assay name in seuratObject. If not provided then default assay in seuratObject will be used.}

\item{multiSamplePattern}{Separator used for multiple samples in seuratObject. Default '_'.}

\item{exprRound}{The expressions are rounded for effectively saving space. Default rounding will be three decimal places.}

\item{spatialSubDir}{Sub-directory name where spatial files are located for 10x.}

\item{sampleInfo}{A dataframe with metadata information for the samples. Rows are for samples.}

\item{downloadRepo}{If TRUE (default), downloads SpatialView from the GitHub repository,
and runs the SpatialView application on a local computer.

When TRUE, the updated SpatialView files are downloaded from GitHub (spatialviewRepo URL).
To download the files, SpatialViewR uses the 'wget' utility. The 'wget' package is pre-installed on most Linux distributions.
For Mac, see \link{https://stackoverflow.com/questions/33886917/how-to-install-wget-in-macos}

For Windows, see   \link{https://gnuwin32.sourceforge.net/packages/wget.htm}

When downloadRepo set to FALSE, SpatialView files may be downloaded manually from the GitHub repository
\link{https://github.com/kendziorski-lab/spatialview/archive/refs/tags/spatialview-latest.zip}
and exportPath to be set to \if{html}{\out{<PATH TO SPATIALVIEW DIR>}}/data/

For details please refer to the \href{https://raw.githubusercontent.com/kendziorski-lab/kendziorski-lab.github.io/main/projects/spatialview/user-guide.pdf}{user-guide}}

\item{spatialviewRepo}{Downloads SpatialView from GitHub repository (ignored if downloadRepo = FALSE).}

\item{spatialviewVersion}{Downloads the specified version of SpatialView (ignored if downloadRepo = FALSE).}

\item{port}{Port to be used for running SpatialView (ignored if downloadRepo = FALSE).}

\item{launchApp}{If TRUE (default), then launches the web application (ignored if downloadRepo = FALSE).}

\item{export_sparse}{If TRUE (default, recommended), sparse expression matrix will be exported, else dense csv format will be used.}

\item{data_file_name_expressions}{Name of the exported expression matrix in csv format (ignored if export_sparse = TRUE).}

\item{data_file_name_expressions_sparse}{Name of the exported expression matrix in sparse format (ignored if export_sparse = FALSE).}

\item{data_file_name_genes}{Name of the exported file containing names of the genes (ignored if export_sparse = FALSE).}

\item{data_file_name_barcodes}{Name of the exported file containing barcodes (ignored if export_sparse = FALSE).}

\item{configList}{A list of named strings for changing the default configuration.

Names should be the names of keys and values should be the corresponding values.

This works if 'downloadRepo' is set to TRUE, else ignored.
Configuration changes are having low precedence as compared to other parameters.}

\item{verbose}{TRUE/FALSE whether to print comments.}
}
\value{
None
}
\description{
Prepares Data from Seurat object and launches SpatialView web application locally.
}
