% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialViewR-package.R
\docType{package}
\name{SpatialViewR-package}
\alias{SpatialViewR}
\alias{SpatialViewR-package}
\title{SpatialViewR: Prepares data for SpatialView}
\description{
This is an helper package for preparing data for using in SpatialView (A web based spatial RNA sequencing data visualization)
}
\author{
\strong{Maintainer}: Chitrasen Mohanty \email{cmohanty2@wisc.edu} (\href{https://orcid.org/0009-0008-0864-6371}{ORCID})

}
\keyword{internal}
